+++ sdk_nnbd/lib/_internal/js_dev_runtime/private/js_helper.dart
@@ -704,25 +704,22 @@ class JSName {
  */
 abstract class JavaScriptIndexingBehavior<E> {}
 
-// TODO(lrn): These exceptions should be implemented in core.
-// When they are, remove the 'Implementation' here.
-
 /// Thrown by type assertions that fail.
 class TypeErrorImpl extends Error implements TypeError {
-  final String message;
+  final String _message;
 
-  TypeErrorImpl(this.message);
+  TypeErrorImpl(this._message);
 
-  String toString() => message;
+  String toString() => _message;
 }
 
 /// Thrown by the 'as' operator if the cast isn't valid.
 class CastErrorImpl extends Error implements CastError {
-  final String message;
+  final String _message;
 
-  CastErrorImpl(this.message);
+  CastErrorImpl(this._message);
 
-  String toString() => message;
+  String toString() => _message;
 }
 
 class FallThroughErrorImplementation extends FallThroughError {