+++ tests/language/assert/assert_test.dart
@@ -8,7 +8,7 @@
 
 import "package:expect/expect.dart";
 
-testTrue() {
+int testTrue() {
   int i = 0;
   try {
     assert(true);
@@ -18,7 +18,7 @@ testTrue() {
   return i;
 }
 
-testFalse() {
+int testFalse() {
   int i = 0;
   try {
     assert(false);
@@ -32,7 +32,7 @@ unknown(dynamic a) {
   return a ? true : false;
 }
 
-testBoolean(bool value) {
+int testBoolean(bool value) {
   int i = 0;
   try {
     assert(value);
@@ -42,7 +42,7 @@ testBoolean(bool value) {
   return i;
 }
 
-testDynamic(dynamic value) {
+int testDynamic(dynamic value) {
   int i = 0;
   try {
     assert(value);
@@ -56,11 +56,10 @@ testMessage(value, message) {
   try {
     assert(value, message);
     return null;
-  } catch (error) {
-    // Catch any type to allow the Boolean conversion to throw either
-    // AssertionError or TypeError.
+  } on AssertionError catch (error) {
     return error;
   }
+  return null;
 }
 
 main() {
@@ -72,20 +71,20 @@ main() {
   Expect.equals(1, testBoolean(false));
   Expect.equals(1, testDynamic(unknown(false)));
 
-  Expect.equals(1, testDynamic(null));
-  Expect.equals(1, testDynamic(42));
-  Expect.equals(1, testDynamic(() => true));
-  Expect.equals(1, testDynamic(() => false));
-  Expect.equals(1, testDynamic(() => 42));
-  Expect.equals(1, testDynamic(() => null));
+  Expect.throwsTypeError(() => testDynamic(null));
+  Expect.throwsTypeError(() => testDynamic(42));
+  Expect.throwsTypeError(() => testDynamic(() => true));
+  Expect.throwsTypeError(() => testDynamic(() => false));
+  Expect.throwsTypeError(() => testDynamic(() => 42));
+  Expect.throwsTypeError(() => testDynamic(() => null));
 
   Expect.equals(1234, testMessage(false, 1234).message);
   Expect.equals('hi', testMessage(false, 'hi').message);
 
-  // These errors do not have the message because boolean conversion failed.
-  Expect.notEquals(1234, testMessage(null, 1234).message);
-  Expect.notEquals('hi', testMessage(null, 'hi').message);
-  Expect.notEquals('hi', testMessage(() => null, 'hi').message);
-  Expect.notEquals('hi', testMessage(() => false, 'hi').message);
-  Expect.notEquals('hi', testMessage(() => true, 'hi').message);
+  // These assertions throw a type error because boolean conversion failed.
+  Expect.throwsTypeError(() => testMessage(null, 1234));
+  Expect.throwsTypeError(() => testMessage(null, 'hi'));
+  Expect.throwsTypeError(() => testMessage(() => null, 'hi'));
+  Expect.throwsTypeError(() => testMessage(() => false, 'hi'));
+  Expect.throwsTypeError(() => testMessage(() => true, 'hi'));
 }